+++ StorageService/PostNew.swift
@@ -1,0 +1,17 @@
+//
+//  PostNew.swift
+//  StorageService
+//
+//  Created by Apple on 31.05.2021.
+//  Copyright © 2021 Artem Novichkov. All rights reserved.
+//
+
+import UIKit
+
+public struct PostNew {
+    public let author: String
+    public let description: String
+    public let image: UIImage
+    public let likes: String
+    public let viewsCount: String
+}
